requirements.txt
Django==5.0.2
djangorestframework==3.14.0
drf-yasg==1.21.7
face-recognition==1.3.0
face-recognition-models==0.3.0
firebase-admin==6.4.0
numpy==1.26.4
pillow==10.2.0
psycopg2-binary==2.9.9
python-decouple==3.8
gunicorn
